Use dynamic page title generator - #108
