+++ .gitmodules
@@ -1,3 +1,4 @@
 [submodule "src/search-ui"]
 	path = src/search-ui
 	url = git@github.com:dbmi-pitt/search-ui.git
+	ignore = all
