+++ src/proxy.js
@@ -1,0 +1,13 @@
+import { NextResponse } from 'next/server';
+
+export function proxy(request) {
+  const requestHeaders = new Headers(request.headers);
+  requestHeaders.set('x-url', request.url);
+  return NextResponse.next({ request: { headers: requestHeaders } });
+}
+
+export const config = {
+  matcher: [
+    '/((?!api|_next/static|_next/image|favicon.ico).*)',
+  ],
+};
